Create compat file for TS 4.6/4.7
Fixes #35
