+++ compat.d.ts
@@ -1,0 +1,2 @@
+export { ErrorWithCause } from "./lib/error-with-cause-compat";
+export * from "./lib/helpers";
