+++ package.json
@@ -10,6 +10,9 @@
   "main": "index.js",
   "module": "index.mjs",
   "types": "index.d.ts",
+  "typesVersions": {
+    "~4.6 || ~4.7": { "index.d.ts": ["compat.d.ts"] }
+  },
   "exports": {
     "require": "./index.js",
     "import": "./index.mjs"
@@ -18,13 +21,18 @@
     "index.js",
     "index.mjs",
     "index.d.ts",
-    "index.d.ts.map"
+    "index.d.ts.map",
+    "lib/**/*.js",
+    "lib/**/*.mjs",
+    "lib/**/*.d.ts",
+    "lib/**/*.d.ts.map"
+
   ],
   "scripts": {
     "build-for-test": "run-s clean build:1:esm",
     "build:0": "run-s clean",
     "build:1:declaration": "tsc -p declaration.tsconfig.json",
-    "build:1:esm": "linemod -e mjs index.js",
+    "build:1:esm": "linemod -e mjs index.js lib/*.js",
     "build:1": "run-p build:1:*",
     "build": "run-s build:*",
     "check:0": "run-s build-for-test",
@@ -36,7 +44,7 @@
     "check:1:type-coverage": "type-coverage --detail --strict --at-least 97 --ignore-files 'test/*'",
     "check:1": "run-p -c --aggregate-output check:1:*",
     "check": "run-s check:*",
-    "clean:declarations": "rm -rf $(find . -maxdepth 2 -type f -name '*.d.ts*')",
+    "clean:declarations": "rm -rf $(find . -maxdepth 2 -type f -name '*.d.ts*' ! -name '*compat.d.ts')",
     "clean": "run-p clean:*",
     "prepare": "husky install",
     "prepublishOnly": "run-s build",
