+++ test-published-types/index.js
@@ -1,0 +1,3 @@
+const { ErrorWithCause } = require('..');
+
+throw new ErrorWithCause('Wow', { cause: new Error('Yay') });
